Added tests for scim admin events

+++ src/test/java/fi/metatavu/keycloak/scim/server/test/tests/functional/RealmUserPatchTestsIT.java
@@ -10,10 +10,14 @@ import fi.metatavu.keycloak.scim.server.test.client.model.PatchRequestOperations
 import fi.metatavu.keycloak.scim.server.test.client.model.User;
 import fi.metatavu.keycloak.scim.server.test.utils.KeycloakTestUtils;
 import org.junit.jupiter.api.Test;
+import org.keycloak.events.admin.AdminEvent;
+import org.keycloak.events.admin.OperationType;
 import org.keycloak.representations.idm.UserRepresentation;
+import org.testcontainers.containers.BindMode;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import java.io.IOException;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -30,6 +34,7 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         .withNetworkAliases("scim-keycloak")
         .withEnv("SCIM_AUTHENTICATION_MODE", "KEYCLOAK")
         .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
+        .withFileSystemBind("build/testdata", "/tmp/testdata", BindMode.READ_WRITE)
         .withRealmImportFile("kc-test.json")
         .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
 
@@ -158,6 +163,46 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         assertEquals("external-5678", patchedAgain.getAdditionalProperty("externalId"));
         assertEquals("Updated Display", patchedAgain.getAdditionalProperty("displayName"));
         assertEquals("en_US", patchedAgain.getAdditionalProperty("preferredLanguage"));
+
+        // Cleanup
+        deleteRealmUser(TestConsts.TEST_REALM, created.getId());
+    }
+
+    @Test
+    void testPatchUserAdminEvents() throws ApiException, IOException {
+        ScimClient scimClient = getAuthenticatedScimClient();
+
+        // Create user
+        User user = new User();
+        user.setUserName("patch-admin-events-user");
+        user.setActive(true);
+        user.setSchemas(List.of("urn:ietf:params:scim:schemas:core:2.0:User"));
+
+        User created = scimClient.createUser(user);
+        clearAdminEvents();
+
+        // Patch user
+        scimClient.patchUser(created.getId(), new PatchRequest()
+                .schemas(List.of("urn:ietf:params:scim:api:messages:2.0:PatchOp"))
+                .operations(List.of(
+                        new PatchRequestOperationsInner()
+                                .op("replace")
+                                .path("userName")
+                                .value("patched-user-name")
+                )));
+
+        List<AdminEvent> adminEvents = getAdminEvents();
+        assertEquals(1, adminEvents.size());
+
+        AdminEvent updateUserEvent = adminEvents.getFirst();
+
+        assertUserAdminEvent(
+            updateUserEvent,
+            TestConsts.TEST_REALM,
+            TestConsts.TEST_REALM_ID,
+            created.getId(),
+            OperationType.UPDATE
+        );
 
         // Cleanup
         deleteRealmUser(TestConsts.TEST_REALM, created.getId());
